Initial Stage
At this point we have a very traditional looking piece of JavaScript.

It is a small library that will fetch some text from the M&S homepage
and it can also submit it to another site.

The code uses a very normal setup for callback functions as well as a
test suite that includes 1 failure. The failure demonstrates a common
mistake people make when using callback based libraries.