Correct failing test
If you have a set of tasks that need to be performed but these are only
performed asynchronously then you need nest them within the predecessors
callback. This is the best way to ensure they only happen when the
necessary steps before it are complete.

The problem we now have though is that this nesting can becoming
difficult to manage and maintain.

So we now have:

TODO: Better asynchronous task chaining. A -> B -> C