Using Promises
Promises provide a means to chaining tasks together, similar to unix and
pipes. The output of the previous step gets passed to the input of the
next step.

What this lets us do is flatten the callback nesting and write very
readable code that works asynchronously.

Things to remember:

then() always takes a function and will convert the result to a promise