Converting arguments
Promises become even more re-usable when you defensively convert
arguments to promises too. This may look like too much initially but it
makes your code more flexible as you can now use values directly or
provided promises that will later resolve the the right value.

I find the best way to think about this is that your code no longer
needs to _know_ the values but it just needs to understand how to get
the values.

This is more of an advanced step but it makes it easier to reuse your
functions and promises in places that you initially hadn't thought they
would be used.